fix: move PWA prompt button to navbar dropdown
